Tidy up rack applications
