+++ apps/web/auth/application.rb
@@ -1,9 +1,18 @@
 # apps/web/auth/application.rb
 
-require 'bundler/setup'
 require 'onetime/application'
 
-require_relative 'app'
+# Load auth dependencies first
+require_relative 'config/database'
+require_relative 'config/rodauth_main'
+require_relative 'helpers/session_validation'
+require_relative 'routes/health'
+require_relative 'routes/validation'
+require_relative 'routes/account'
+require_relative 'routes/admin'
+
+# Load Roda app
+require_relative 'router'
 
 module Auth
   class Application < Onetime::Application::Base
@@ -11,21 +20,15 @@ module Auth
 
     # Auth app specific middleware (common middleware is in MiddlewareStack)
 
-    # Development Environment Configuration
-    # Enable development-specific middleware when in development mode
-    # This handles code validation and frontend development server integration
     Onetime.development? do
-
+      # Development configuration if needed
     end
 
-    # Serve static frontend assets in production mode
-    # While reverse proxies often handle static files in production,
-    # this provides a fallback capability for simpler deployments.
     Onetime.production? do
       # Production configuration
       use Rack::Deflater  # Gzip compression
 
-      # Security headers
+      # Security headers (some may be redundant with MiddlewareStack)
       use Rack::Protection::AuthenticityToken
       use Rack::Protection::ContentSecurityPolicy
       use Rack::Protection::FrameOptions
@@ -37,17 +40,15 @@ module Auth
     end
 
     warmup do
-      # Expensive initialization tasks go here
-
-      # Log warmup completion
-      Onetime.li 'Auth warmup completed'
+      # Dependencies already loaded, just warmup tasks here
     end
 
     protected
 
     def build_router
-      # Return the Roda app instance directly (not frozen)
-      AuthService.app
+      # Return the Roda app instance
+      # Unlike Otto apps, Roda apps are classes that respond to call
+      Auth::Router
     end
   end
 end
