+++ lib/onetime/middleware.rb
@@ -7,3 +7,4 @@ require_relative '../middleware/detect_host'
 require_relative 'middleware/domain_strategy'
 require_relative 'middleware/identity_resolution'
 require_relative 'middleware/startup_readiness'
+require_relative 'middleware/security'
